alloc_slab: change allocator from heap of size bins to bitmap
size bins are very fast O(1), however they are not dense as
chunks can't be subsequently split up when size demand changes.
thus we change alloc_slab to use a simple bitmap allocator.

note: this initial bitmap allocator is naive. it exhastively
scans to find the first fit. a subsequent revision could add
statistics to perform stochastic best fit by recording size
statistics and being tactical about which bitmap chunks are
split. we could also partition the slab based on page index
to reduce the worst case scan time at the expense of requiring
a rebalancing of the slab partitions when the slab is resized.
we would keep slab partition balanced and use the heuristic of
a partition full condition to trigger slab resize.